+++ matrix.hpp
@@ -29,15 +29,15 @@
 #include "build.hpp"
 namespace filter::matrix {
     // function prototypes
-    void f3x3matrixAeqI(float A[][3]);
-    void fmatrixAeqI(float* A[], int16 rc);
-    void f3x3matrixAeqScalar(float A[][3], float Scalar);
-    void f3x3matrixAeqInvSymB(float A[][3], float B[][3]);
-    void f3x3matrixAeqAxScalar(float A[][3], float Scalar);
-    void f3x3matrixAeqMinusA(float A[][3]);
-    float f3x3matrixDetA(float A[][3]);
-    void eigencompute(float A[][10], float eigval[], float eigvec[][10], int8 n);
-    void fmatrixAeqInvA(float* A[], int8 iColInd[], int8 iRowInd[], int8 iPivot[], int8 isize);
-    void fmatrixAeqRenormRotA(float A[][3]);
+    void f3x3matrixAeqI(double A[][3]);
+    void fmatrixAeqI(double* A[], int16 rc);
+    void f3x3matrixAeqScalar(double A[][3], double Scalar);
+    void f3x3matrixAeqInvSymB(double A[][3], double B[][3]);
+    void f3x3matrixAeqAxScalar(double A[][3], double Scalar);
+    void f3x3matrixAeqMinusA(double A[][3]);
+    double f3x3matrixDetA(double A[][3]);
+    void eigencompute(double A[][10], double eigval[], double eigvec[][10], int8 n);
+    void fmatrixAeqInvA(double* A[], int8 iColInd[], int8 iRowInd[], int8 iPivot[], int8 isize);
+    void fmatrixAeqRenormRotA(double A[][3]);
 }  // namespace filter::matrix
 #endif  // #ifndef MATRIX_HPP
